added pagerank algo

+++ dijkstras.h
@@ -5,7 +5,7 @@
 
 class dijkstras {
     public:
-        std::vector<float> dijkstra_path(std::vector<std::vector<int>> graph, int start);
+        std::vector<float> dijkstra_path(std::vector<std::vector<float>> graph, int start);
     private:
         int minDistance(std::vector<float> distances, std::vector<int> shortest_path);
 };
